Refactoring Syntax and Instances for XmlZoom, NestedMap, Predicate.
