--- src/main/scala/advxml/syntax/NestedMapSyntax.scala
@@ -1,14 +1,0 @@
-package advxml.syntax
-
-import cats.{Applicative, Monad}
-import cats.implicits._
-
-private[syntax] trait NestedMapSyntax {
-  implicit class ApplicativeDeepMapOps[F[_]: Applicative, G[_]: Applicative, A](fg: F[G[A]]) {
-    def nestedMap[B](f: A => B): F[G[B]] = fg.map(_.map(f))
-  }
-
-  implicit class ApplicativeDeepFlatMapOps[F[_]: Applicative, G[_]: Monad, A](fg: F[G[A]]) {
-    def nestedFlatMap[B](f: A => G[B]): F[G[B]] = fg.map(_.flatMap(f))
-  }
-}
