--- src/main/scala/advxml/syntax/PredicateSyntax.scala
@@ -1,45 +1,0 @@
-package advxml.syntax
-
-import advxml.core.Predicate
-
-private[syntax] trait PredicateSyntax {
-
-  implicit class PredicateOps[T](p: T => Boolean) {
-
-    /** Combine with another predicate(`T => Boolean`) with `And` operator.
-      *
-      * @see [[Predicate]] object for further information.
-      * @param that Predicate to combine.
-      * @return Result of combination with this instance
-      *         with passed predicate instance using `And` operator.
-      */
-    def &&(that: T => Boolean): T => Boolean = p.and(that)
-
-    /** Combine with another predicate(`T => Boolean`) with `And` operator.
-      *
-      * @see [[Predicate]] object for further information.
-      * @param that Predicate to combine.
-      * @return Result of combination with this instance
-      *         with passed predicate instance using `And` operator.
-      */
-    def and(that: T => Boolean): T => Boolean = Predicate.and(p, that)
-
-    /** Combine with another predicate(`T => Boolean`) with `Or` operator.
-      *
-      * @see [[Predicate]] object for further information.
-      * @param that Predicate to combine.
-      * @return Result of combination with this instance
-      *         with passed predicate instance using `Or` operator.
-      */
-    def ||(that: T => Boolean): T => Boolean = p.or(that)
-
-    /** Combine with another predicate(`T => Boolean`) with `Or` operator.
-      *
-      * @see [[Predicate]] object for further information.
-      * @param that Predicate to combine.
-      * @return Result of combination with this instance
-      *         with passed predicate instance using `Or` operator.
-      */
-    def or(that: T => Boolean): T => Boolean = Predicate.or(p, that)
-  }
-}
